+++ fixtures/path-to-libraries/source/javascripts/application.js
@@ -1,3 +1,2 @@
-//=require handlebars
 //=require ember
 //=require ember-data